builds now
ShowViewController.swift has issues. I am trying to figure out how to
link the buttons so that the associated website is opened when the
button is pressed.

Additionally, there is a layout issue where the buttons aren’t fit into
the screen layout (scroll down to see buttons if they are hiding)