Support DJ bios (#30)
* Added functionality to display DJ bios

Implemented pod 'Presentr' to create a popup display of a new DJPopupViewController which displays the DJ name, DJ real name, and DJ biography.  Had to modify DJ.swift to support the retrieval of DJ bios.

* Forgot to update readme earlier!  Also fixed an issue where the DJ bio would automatically scroll to the bottom if there was enough text.  Text now starts at the top.

* Stop tracking pods

* Resubmitting branch /support-dj-bios on new branch brian/support-dj-bios

* readme presentr was formatted incorrectly.

* Fixed bug where the image isn't cropped when you background app and go back.  Bear image cropping is still odd.

* Fixed bugs associated with DJ images

Bugs included not resizing images when backgrounding app and returning to app.  And bear would not resize correctly.

* Synced up